+++ CHANGELOG.md
@@ -2,14 +2,21 @@
 
 All notable changes to this project will be documented in this file.
 
-## 0.1.7 Feb 18, 2021
+## [0.2.0] - 2021-02-24
+
+### Added
+
+- TONOS-SE Controller.
+
+## [0.1.7] – 2021-02-18
+
 ### New
 
-- Solidity functionality added: 
+- Solidity functionality added:
   - Create a hello world contract in one click
   - Install compiler in the background upon the first request
   - Compile and link a contract in one click
-  - Get solidity compiler version 
+  - Get solidity compiler version
   - Download the latest compiler
 - SDK functionality added:
   - Create a Node.js project with the latest SDK dependencies and index.js script with main client object creation